Revamping the recipe site
After continued research I found a video that explained how to use css and html only to develop a site. To remaster css and html I want to practice building the bones of site with the two, while incorperating flex box (later on)

Today I changed the layout of the home page and added the resources I used to help do so.

Tonight I will be rebuilding the css stylesheet and applying it to each page.