The Home Page Is Done!
I was able to complete the layout and design of the home page.

I began working on the Mac & Cheese page by creating a stylesheet and styling the navbar

Tonight I will continue the edits and finish the Mac page
I will also begin work on the PB & J page (the stylesheet has been created for this, I will need to link it for the navbar to reflect the changes).