Final Edits to Gihub Pages
I struggled a bit with the delay in the pages (updates)
It confused me for a bit, I thought my code was wrong
I finally looked up the issue and saw that updates can take up to ten minutes to show. I will make this last edit and allow the pages to reflect the changes after 15 minutes.

This is a major downside in puplishing your sites on github pages, but once you know it you are good to go.